[NRRPLT-2797][NRRPLT-2794] Support specifying animated robot model for gz3d scene.
Adds support to propagate the animated model specification when
loading a new gz3d scene. This model is specified in the ExDConf
sent from the backend with other experiment data and will replace
any server side robot meshes (if provided).

This mechanism works in the same way as specifying the default
camera position in a scene when it is loaded (which is also
specified in the ExDConf and propagated here by the backend with
the experiment details).

Propagates gz3d.js changes made in gzweb for the same issue
[NRRPLT-2797].

Added additional unrelated test case coverage to pass line threshold,
which failed after rebasing against master.

Change-Id: I21851ada2503780eff72f18961ac66dab8f01e73