[hotifx] Add local install config.json to .gitignore.
app/config.json.local is downloaded from the wiki as part of the
local install and should be ignored (and not accidentally committed
to git) as is already done with app/config.json (users are instructed
to create a simlink config.json -> config.json.local during the local
install process).

The rationale behind the simlink is to allow users to switch between
config.json.local and config.json.sample to run locally vs deployed.

Change-Id: I3c9d53d88a9810746a9baaba289f3122372f6e70