[NRRPLT-4190] Fix initial camera/animated model race condition.
A rare race condition existed where the animated model and initial
camera position would try to be set before the gz3d scene was available.
The calls would then be ignored and the expeirment would load with the
wrong camera position and an invisible mouse in the mouse experiments
(because they are the only experiments to use an animated mesh).

The calls must be made after the gz3d scene is created, but cannot be
delayed until after the entire scene is initialized because the call to
set the animated model must occur before gz3d attempts to load all models
in the scene. This is because of the logic in gz3d that required the animated
model to be set so that it only loads the animated model rather than each
individual mouse mesh.

Change-Id: I4425fc3e7f461468b6f6c11246726f58ea7197d0