[HOTFIX] Refers only to latest online documentation.
Instruction for testing: open the Transfer function editor, the SMACH script editor and the Brain editor and click on the documentation links.
You should be redirected to an url of the following  forms:
- https://developer.humanbrainproject.eu/docs/hbp-nrp-cle/latest/*
- https://developer.humanbrainproject.eu/docs/hbp_nrp_backend/latest/*
- https://developer.humanbrainproject.eu/docs/HBP Neurorobotics Platform/latest/*

When the commit https://bbpcode.epfl.ch/code/#/c/29599/ is deployed, you can check the redirect links are actually correct.

Change-Id: I6a32441a56d523eff13adb8b1097510dbe6d5ca6