[NRRPLT-4190] High latency camera/animated model race condition fix.
This patch is a second fix for the invisible mouse / misplaced camera bug
that occured during testing on a high latency connection (from the US to
the dev servers).

Unfortunately angular.isDefined(...) will return true if a variable is
undefined, which led to a very particular timing issue on a slow network
where scene would be defined but set to undefined. This patch resolves
the issue by properly checking for valid/defined scene and removing the
checks that were allowing for camera/animated model setting to fail silently.

Change-Id: Ib29beb0c4d9b50da07c199d9639cbcb2c76a48e0