[NRRPLT-4544] Disable brain reset for multi-process brains.
This patch disables the user options to reset the brain (and the
full reset option) for multi-process distributed brain simulations
as this is not supported by the backend. It removes the options from
the reset menu and displays text to indicate that they are currently
disabled.

The patch also fixes an issue with the "Launch in Single Process"
button that was not behaving the same as the normal Launch button and
being properly enabled/disabled after the user clicked one of them.

Change-Id: Iff7a5bf378b28c59719a218b49f89787081fe595