[NRRPLT-4515][NRRPLT-4516][NRRPLT-4518] Shape/cloud support in the brainvisualizer
This includes a new version of the brainvisualizer which supports
several new shape representations:
- Sphere
- Cube
- Flat
- Cloud

For each representation you can choose how the neurons are distributed:
- Overlap: Neurons owned by several populations overlap

- Distribute: All neurons are splited by pop.  and distributed on the
shape

- Split: Populations are splitted in several overlapping shape

To test with a high number of neurons, you can go to the brain
editor and changes the populations. Try with hundreds, then thousands,
etc. I was able to run it with one million neurons on my machine
on chrome.

Change-Id: Idb3a9c12100c7dad17c0603d4472bcd18ae31da4