[NRRPLT-4729][NRRPLT-4743] Simulation status unsubscribe race condition.
This patch fixes a race condition that appears when running a
multi-process brain simulation that involved the simulation/CLE
status topic.

The root of the issue is that roslib.js calls to unsubscribe fully
disconnect the rosbrige from a given topic, rather than just
removing the inidivudal listener. Further, passing the callback to
unsubscribe does not fix the issue if there are multiple instances
of roslib Topics in different contexts - if you unregister a listener
in one instance and it has no more listeners then it will fully
disconnect the rosbridge topic even if there are other instances for
the same topic with active listeners.

The specific issue was between the status/loading progress bar and
the main gz3d view. When running a multi-process brain simulation,
the loading order/timing was slightly different so that gz3d registered
for status messages before the progress bar called unsubscribe. This
meant that the gz3d status connection was fully terminated - causing
the bug behavior of static time displays and not properly terminating
the simulation (which checks for a stopped status message to transition
to the shutdown splash screen).

This doesn't seem to be an issue anywhere else as the status topic
seems to be the only shared topic that has multiple purposes across
a simulation lifecycle.

Change-Id: I5d980a229e3681983581cafab23a43929eca2c85