[NRRPLT-4891] Disable TF/brain update functionality for distributed simulations.
This patch disables all functionality to add/delete/update transfer
functions or the brain when the user is running a distributed simulation
as the backend cannot currently support that behavior.

It also disables the display of the "population" views, as they do not
currently make sense due to the way MUSIC port/population configuration
is done.

Change-Id: I3310b268d30caa036b72f902ee3a7000152f999e