[NRRPLT-5654] Repair luminosity +/-
This repairs the luminosity feature in the first place. This means, the
feature is restored exactly in the way it was before, so we do not
set ambient lights nor do we change light intensities above 1 or below 0.

How to test: Open a new experiment and play with the light. Note that it has
an effect on both the client stream and the images rendered on the server.

Change-Id: Iad9ecce67ae390bd8cf98881f7b6118cd33af7ba